requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.18

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Export functionality
openpyxl>=3.1.2
reportlab>=4.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Testing (optional)
pytest>=7.4.0
httpx>=0.26.0

